Add files via upload
This was one of my first CSS project where we had to copy the original photo exactly. This was my first time being introduced to flex-box and understanding it more in depth. The most time consuming was getting the exact sizing of all the pictures and sections but with the help of some of my fellow classmates, I was able to complete it.